% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{parse_boulder}
\alias{parse_boulder}
\alias{read_boulder}
\title{Read Boulder-IO formatted files.}
\usage{
parse_boulder(strings, single = FALSE)

read_boulder(file, ...)
}
\arguments{
\item{strings}{Character vector of lines from a Boulder-IO file.}

\item{single}{By default, a Boulder-IO file contains multiple records, so a list of
records is returned. If \code{single=TRUE} only a single record object (the first) is
returned.}

\item{file}{File connection object or filename.}
}
\value{
A list of records. Each record is also a list of tags (names) and values. If
The Boulder-IO record contained multiple lines with the same tag, the values are
joined into a single vector in the list. All values are character strings.
}
\description{
Read Boulder-IO formatted files.
}
\section{Functions}{
\itemize{
\item \code{parse_boulder}: parses strings in the Boulder-IO format, with each string representing a line.

\item \code{read_boulder}: reads a text file containing
one or more Boulder-IO records.
}}

