% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{write_boulder}
\alias{write_boulder}
\title{Write a list of records to a Boulder-IO text file.}
\usage{
write_boulder(records, file = "", single = FALSE)
}
\arguments{
\item{records}{A list of records. Each record is also a list of tags (names) and
values. All values are converted with \code{as.character}. Tags with a vector of
values are written on separate lines with the same tag.}

\item{file}{A filename or a connection object. Use "" for STDOUT.}

\item{single}{If \code{TRUE}, the input \code{records} is only a single record,
not a list of records.}
}
\description{
Write a list of records to a Boulder-IO text file.
}
